general advancement

+++ src/entities/point.hpp
@@ -9,6 +9,6 @@
 #ifndef POINT_HPP
 #define POINT_HPP
 
-struct Point { float x_, y_; };
+struct Point { double x_, y_; };
 
 #endif
